# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/services": {
      "post": {
        "operationId": "ServicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Services criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Criar Services",
        "tags": [
          "Services"
        ]
      },
      "get": {
        "operationId": "ServicesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "complexity",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Listar todos Services",
        "tags": [
          "Services"
        ]
      }
    },
    "/services/export": {
      "get": {
        "operationId": "ServicesController_exportCsv",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation exportCsv",
        "tags": [
          "Services"
        ]
      }
    },
    "/services/import/validate": {
      "post": {
        "operationId": "ServicesController_validateCsv",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Services criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation validateCsv",
        "tags": [
          "Services"
        ]
      }
    },
    "/services/import/confirm": {
      "post": {
        "operationId": "ServicesController_confirmImport",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Services criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation confirmImport",
        "tags": [
          "Services"
        ]
      }
    },
    "/services/{id}": {
      "get": {
        "operationId": "ServicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Buscar um Services",
        "tags": [
          "Services"
        ]
      },
      "put": {
        "operationId": "ServicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Atualizar Services",
        "tags": [
          "Services"
        ]
      },
      "delete": {
        "operationId": "ServicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remover Services",
        "tags": [
          "Services"
        ]
      }
    },
...
    "/technicians": {
      "post": {
        "operationId": "TechniciansController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTechnicianDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Technicians criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Criar Technicians",
        "tags": [
          "Technicians"
        ]
      },
      "get": {
        "operationId": "TechniciansController_findAll",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Listar todos Technicians",
        "tags": [
          "Technicians"
        ]
      }
    },
    "/technicians/ranking": {
      "get": {
        "operationId": "TechniciansController_getRanking",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation getRanking",
        "tags": [
          "Technicians"
        ]
      }
    },
    "/technicians/{id}": {
      "get": {
        "operationId": "TechniciansController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Buscar um Technicians",
        "tags": [
          "Technicians"
        ]
      },
      "put": {
        "operationId": "TechniciansController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTechnicianDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Atualizar Technicians",
        "tags": [
          "Technicians"
        ]
      },
      "delete": {
        "operationId": "TechniciansController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remover Technicians",
        "tags": [
          "Technicians"
        ]
      }
    },
...
  },
  "info": {
    "title": "Click Marido API",
    "description": "API documentation for Click Marido ERP/CRM",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "description": "Campo category",
            "example": "exemplo"
          },
          "name": {
            "type": "string",
            "description": "Campo name",
            "example": "exemplo"
          },
          "description": {
            "type": "string",
            "description": "Campo description",
            "example": "exemplo"
          },
          "value": {
            "type": "number",
            "description": "Campo value",
            "example": 1
          },
          "averageTime": {
            "type": "number",
            "description": "Campo averageTime",
            "example": 1
          },
          "complexity": {
            "type": "string",
            "description": "Campo complexity",
            "example": "exemplo"
          },
          "warranty": {
            "type": "number",
            "description": "Campo warranty",
            "example": 1
          },
          "specialty": {
            "type": "string",
            "description": "Campo specialty",
            "example": "exemplo"
          },
          "active": {
            "type": "boolean",
            "description": "Campo active",
            "example": true
          }
        },
        "required": [
          "category",
          "name",
          "value",
          "averageTime",
          "complexity",
          "warranty"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "description": "Campo category",
            "example": "exemplo"
          },
          "name": {
            "type": "string",
            "description": "Campo name",
            "example": "exemplo"
          },
          "description": {
            "type": "string",
            "description": "Campo description",
            "example": "exemplo"
          },
          "value": {
            "type": "number",
            "description": "Campo value",
            "example": 1
          },
          "averageTime": {
            "type": "number",
            "description": "Campo averageTime",
            "example": 1
          },
          "complexity": {
            "type": "string",
            "description": "Campo complexity",
            "example": "exemplo"
          },
          "warranty": {
            "type": "number",
            "description": "Campo warranty",
            "example": 1
          },
          "specialty": {
            "type": "string",
            "description": "Campo specialty",
            "example": "exemplo"
          },
          "active": {
            "type": "boolean",
            "description": "Campo active",
            "example": true
          }
        }
      },
...
      "CreateTechnicianDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Campo name",
            "example": "exemplo"
          },
          "phone": {
            "type": "string",
            "description": "Campo phone",
            "example": "exemplo"
          },
          "specialty": {
            "type": "string",
            "description": "Campo specialty",
            "example": "exemplo"
          },
          "rating": {
            "type": "number",
            "description": "Campo rating",
            "example": 1
          },
          "status": {
            "type": "string",
            "description": "Campo status",
            "example": "exemplo"
          },
          "companyId": {
            "type": "string",
            "description": "Campo companyId",
            "example": "exemplo"
          }
        },
        "required": [
          "name",
          "phone",
          "companyId"
        ]
      },
      "UpdateTechnicianDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Campo name",
            "example": "exemplo"
          },
          "phone": {
            "type": "string",
            "description": "Campo phone",
            "example": "exemplo"
          },
          "specialty": {
            "type": "string",
            "description": "Campo specialty",
            "example": "exemplo"
          },
          "rating": {
            "type": "number",
            "description": "Campo rating",
            "example": 1
          },
          "status": {
            "type": "string",
            "description": "Campo status",
            "example": "exemplo"
          }
        }
      },
...
    }
  }
}